phc_t1_freesurfer

NACCADRC ADSP PHC Imaging T1 dataset. T1 Freesurfer Harmonization Data File. The data is sourced from the file NACC_ADSP_PHC_T1_Freesurfer_2024.csv.

Usage

data(phc_t1_freesurfer)

Format

A data frame with 7740 rows and 194 variables:

NACCID

Participant ID

NACCVNUM

Visit Number

X3rd.Ventricle_combat

3rd ventricle volume harmonized with combat

X4th.Ventricle_combat

4th ventricle volume harmonized with combat

Brain.Stem_combat

Brainstem volume harmonized with combat

BrainSegVol.to.eTIV_combat

Ratio of total brain segmentation volume to ICV harmonized with combat

BrainSegVol_combat

Segmented brain volume harmonized with combat

BrainSegVolNotVentSurf_combat

Surface size of the segmented brain volume excluding ventricles harmonized with combat

CC_Anterior_combat

Anterior corpus collosum volume harmonized with combat

CC_Central_combat

Central corpus collosum volume harmonized with combat

CC_Mid_Anterior_combat

Mid-anterior corpus collosum volume harmonized with combat

CC_Mid_Posterior_combat

Mid-posterior corpus collosum volume harmonized with combat

CC_Posterior_combat

Posterior corpus collosum volume harmonized with combat

CerebralWhiteMatterVol_combat

Total cerebral cortex white matter volume harmonized with combat

CortexVol_combat

Total cerebral cortex grey matter volume harmonized with combat

CSF_combat

CSF volume harmonized with combat

EstimatedTotalIntraCranialVol_combat

Total intracranial volume harmonized with combat

Left.Accumbens.area_combat

Left accumbens volume harmonized with combat

Left.Amygdala_combat

Left amygdala volume harmonized with combat

Left.Caudate_combat

Left caudate volume harmonized with combat

Left.Cerebellum.Cortex_combat

Left cerebellar cortex volume harmonized with combat

Left.Cerebellum.White.Matter_combat

Left cerebellar white matter volume harmonized with combat

Left.choroid.plexus_combat

Left choroid plexus volume harmonized with combat

Left.Hippocampus_combat

Left hippocampus volume harmonized with combat

Left.Inf.Lat.Vent_combat

Left inferior lateral ventricle volume harmonized with combat

Left.Lateral.Ventricle_combat

Left lateral ventricle volume harmonized with combat

Left.Pallidum_combat

Left pallidum volume harmonized with combat

Left.Putamen_combat

Left putamen volume harmonized with combat

Left.Thalamus.Proper_combat

Left thalamus volume harmonized with combat

Left.VentralDC_combat

Left ventral DC volume harmonized with combat

lh_bankssts_thickness_combat

Left banks of the superior temporal sulcus thickness harmonized with combat

lh_bankssts_volume_combat

Left banks of the superior temporal sulcus grey matter volume harmonized with combat

lh_caudalanteriorcingulate_thickness_combat

Left caudal anterior cingulate thickness harmonized with combat

lh_caudalanteriorcingulate_volume_combat

Left caudal anterior cingulate grey matter volume harmonized with combat

lh_caudalmiddlefrontal_thickness_combat

Left caudal middle frontal gyri thickness harmonized with combat

lh_caudalmiddlefrontal_volume_combat

Left caudal middle frontal gyri grey matter volume harmonized with combat

lh_cuneus_thickness_combat

Left cuneus thickness harmonized with combat

lh_cuneus_volume_combat

Left cuneus grey matter volume harmonized with combat

lh_entorhinal_thickness_combat

Left entorhinal cortex thickness harmonized with combat

lh_entorhinal_volume_combat

Left entorhinal cortex grey matter volume harmonized with combat

lh_frontalpole_thickness_combat

Left frontal pole thickness harmonized with combat

lh_frontalpole_volume_combat

Left frontal pole grey matter volume harmonized with combat

lh_fusiform_thickness_combat

Left fusiform gyri thickness harmonized with combat

lh_fusiform_volume_combat

Left fusiform gyri grey matter volume harmonized with combat

lh_inferiorparietal_thickness_combat

Left inferior parietal gyri thickness harmonized with combat

lh_inferiorparietal_volume_combat

Left inferior parietal gyri grey matter volume harmonized with combat

lh_inferiortemporal_thickness_combat

Left inferior temporal gyri thickness harmonized with combat

lh_inferiortemporal_volume_combat

Left inferior temporal gyri grey matter volume harmonized with combat

lh_insula_thickness_combat

Left insula thickness harmonized with combat

lh_insula_volume_combat

Left insula grey matter volume harmonized with combat

lh_isthmuscingulate_thickness_combat

Left isthmus cingulate thickness harmonized with combat

lh_isthmuscingulate_volume_combat

Left isthmus cingulate grey matter volume harmonized with combat

lh_lateraloccipital_thickness_combat

Left lateral occipital gyri thickness harmonized with combat

lh_lateraloccipital_volume_combat

Left lateral occipital gyri grey matter volume harmonized with combat

lh_lateralorbitofrontal_thickness_combat

Left lateral orbitofrontal gyri thickness harmonized with combat

lh_lateralorbitofrontal_volume_combat

Left lateral orbitofrontal gyri grey matter volume harmonized with combat

lh_lingual_thickness_combat

Left lingual gyri thickness harmonized with combat

lh_lingual_volume_combat

Left lingual gyri grey matter volume harmonized with combat

lh_MeanThickness_thickness_combat

Left mean lobar thickness harmonized with combat

lh_medialorbitofrontal_thickness_combat

Left medial orbitofrontal thickness harmonized with combat

lh_medialorbitofrontal_volume_combat

Left medial orbitofrontal grey matter volume harmonized with combat

lh_middletemporal_thickness_combat

Left middle temporal gyri thickness harmonized with combat

lh_middletemporal_volume_combat

Left middle temporal gyri grey matter volume harmonized with combat

lh_paracentral_thickness_combat

Left paracentral lobule thickness harmonized with combat

lh_paracentral_volume_combat

Left paracentral lobule grey matter volume harmonized with combat

lh_parahippocampal_thickness_combat

Left parahippocampal gyri thickness harmonized with combat

lh_parahippocampal_volume_combat

Left parahippocampal gyri grey matter volume harmonized with combat

lh_parsopercularis_thickness_combat

Left pars opercularis thickness harmonized with combat

lh_parsopercularis_volume_combat

Left pars opercularis grey matter volume harmonized with combat

lh_parsorbitalis_thickness_combat

Left pars orbitalis thickness harmonized with combat

lh_parsorbitalis_volume_combat

Left pars orbitalis grey matter volume harmonized with combat

lh_parstriangularis_thickness_combat

Left pars triangularis thickness harmonized with combat

lh_parstriangularis_volume_combat

Left pars triangularis grey matter volume harmonized with combat

lh_pericalcarine_thickness_combat

Left pericalcarine gyri thickness harmonized with combat

lh_pericalcarine_volume_combat

Left pericalcarine gyri grey matter volume harmonized with combat

lh_postcentral_thickness_combat

Left postcentral gyri thickness harmonized with combat

lh_postcentral_volume_combat

Left postcentral gyri grey matter volume harmonized with combat

lh_posteriorcingulate_thickness_combat

Left posterior cingulate thickness harmonized with combat

lh_posteriorcingulate_volume_combat

Left posterior cingulate grey matter volume harmonized with combat

lh_precentral_thickness_combat

Left precentral gyri thickness harmonized with combat

lh_precentral_volume_combat

Left precentral gyri grey matter volume harmonized with combat

lh_precuneus_thickness_combat

Left precuneus thickness harmonized with combat

lh_precuneus_volume_combat

Left precuneus grey matter volume harmonized with combat

lh_rostralanteriorcingulate_thickness_combat

Left rostral anterior cingulate thickness harmonized with combat

lh_rostralanteriorcingulate_volume_combat

Left rostral anterior cingulate grey matter volume harmonized with combat

lh_rostralmiddlefrontal_thickness_combat

Left rostral middle frontal gyri thickness harmonized with combat

lh_rostralmiddlefrontal_volume_combat

Left rostral middle frontal gyri grey matter volume harmonized with combat

lh_superiorfrontal_thickness_combat

Left superior frontal gyri thickness harmonized with combat

lh_superiorfrontal_volume_combat

Left superior frontal gyri grey matter volume harmonized with combat

lh_superiorparietal_thickness_combat

Left superior parietal gyri thickness harmonized with combat

lh_superiorparietal_volume_combat

Left superior parietal gyri grey matter volume harmonized with combat

lh_superiortemporal_thickness_combat

Left superior temporal gyri thickness harmonized with combat

lh_superiortemporal_volume_combat

Left superior temporal gyri grey matter volume harmonized with combat

lh_supramarginal_thickness_combat

Left supramarginal gyri thickness harmonized with combat

lh_supramarginal_volume_combat

Left supramarginal gyri grey matter volume harmonized with combat

lh_temporalpole_thickness_combat

Left temporal pole thickness harmonized with combat

lh_temporalpole_volume_combat

Left temporal pole grey matter volume harmonized with combat

lh_transversetemporal_thickness_combat

Left transverse temporal pole thickness harmonized with combat

lh_transversetemporal_volume_combat

Left transverse temporal pole grey matter volume harmonized with combat

lhCerebralWhiteMatterVol_combat

Left cerebral cortex white matter volume harmonized with combat

lhCortexVol_combat

Left cerebral cortex grey matter volume harmonized with combat

MaskVol.to.eTIV_combat

Ratio of mask segmentation volume to ICV harmonized with combat

MaskVol_combat

Mask volume harmonized with combat

Optic.Chiasm_combat

Optic Chiasm volume harmonized with combat

rh_bankssts_thickness_combat

Right banks of the superior temporal sulcus thickness harmonized with combat

rh_bankssts_volume_combat

Right banks of the superior temporal sulcus grey matter volume harmonized with combat

rh_caudalanteriorcingulate_thickness_combat

Right caudal anterior cingulate thickness harmonized with combat

rh_caudalanteriorcingulate_volume_combat

Right caudal anterior cingulate grey matter volume harmonized with combat

rh_caudalmiddlefrontal_thickness_combat

Right caudal middle frontal gyri thickness harmonized with combat

rh_caudalmiddlefrontal_volume_combat

Right caudal middle frontal gyri grey matter volume harmonized with combat

rh_cuneus_thickness_combat

Right cuneus thickness harmonized with combat

rh_cuneus_volume_combat

Right cuneus grey matter volume harmonized with combat

rh_entorhinal_thickness_combat

Right entorhinal cortex thickness harmonized with combat

rh_entorhinal_volume_combat

Right entorhinal cortex grey matter volume harmonized with combat

rh_frontalpole_thickness_combat

Right frontal pole thickness harmonized with combat

rh_frontalpole_volume_combat

Right frontal pole grey matter volume harmonized with combat

rh_fusiform_thickness_combat

Right fusiform gyri thickness harmonized with combat

rh_fusiform_volume_combat

Right fusiform gyri grey matter volume harmonized with combat

rh_inferiorparietal_thickness_combat

Right inferior parietal gyri thickness harmonized with combat

rh_inferiorparietal_volume_combat

Right inferior parietal gyri grey matter volume harmonized with combat

rh_inferiortemporal_thickness_combat

Right inferior temporal gyri thickness harmonized with combat

rh_inferiortemporal_volume_combat

Right inferior temporal gyri grey matter volume harmonized with combat

rh_insula_thickness_combat

Right insula thickness harmonized with combat

rh_insula_volume_combat

Right insula grey matter volume harmonized with combat

rh_isthmuscingulate_thickness_combat

Right isthmus cingulate thickness harmonized with combat

rh_isthmuscingulate_volume_combat

Right isthmus cingulate grey matter volume harmonized with combat

rh_lateraloccipital_thickness_combat

Right lateral occipital gyri thickness harmonized with combat

rh_lateraloccipital_volume_combat

Right lateral occipital gyri grey matter volume harmonized with combat

rh_lateralorbitofrontal_thickness_combat

Right lateral orbitofrontal gyri thickness harmonized with combat

rh_lateralorbitofrontal_volume_combat

Right lateral orbitofrontal gyri grey matter volume harmonized with combat

rh_lingual_thickness_combat

Right lingual gyri thickness harmonized with combat

rh_lingual_volume_combat

Right lingual gyri grey matter volume harmonized with combat

rh_MeanThickness_thickness_combat

Right mean lobar thickness harmonized with combat

rh_medialorbitofrontal_thickness_combat

Right medial orbitofrontal thickness harmonized with combat

rh_medialorbitofrontal_volume_combat

Right medial orbitofrontal grey matter volume harmonized with combat

rh_middletemporal_thickness_combat

Right middle temporal gyri thickness harmonized with combat

rh_middletemporal_volume_combat

Right middle temporal gyri grey matter volume harmonized with combat

rh_paracentral_thickness_combat

Right paracentral lobule thickness harmonized with combat

rh_paracentral_volume_combat

Right paracentral lobule grey matter volume harmonized with combat

rh_parahippocampal_thickness_combat

Right parahippocampal gyri thickness harmonized with combat

rh_parahippocampal_volume_combat

Right parahippocampal gyri grey matter volume harmonized with combat

rh_parsopercularis_thickness_combat

Right pars opercularis thickness harmonized with combat

rh_parsopercularis_volume_combat

Right pars opercularis grey matter volume harmonized with combat

rh_parsorbitalis_thickness_combat

Right pars orbitalis thickness harmonized with combat

rh_parsorbitalis_volume_combat

Right pars orbitalis grey matter volume harmonized with combat

rh_parstriangularis_thickness_combat

Right pars triangularis thickness harmonized with combat

rh_parstriangularis_volume_combat

Right pars triangularis grey matter volume harmonized with combat

rh_pericalcarine_thickness_combat

Right pericalcarine gyri thickness harmonized with combat

rh_pericalcarine_volume_combat

Right pericalcarine gyri grey matter volume harmonized with combat

rh_postcentral_thickness_combat

Right postcentral gyri thickness harmonized with combat

rh_postcentral_volume_combat

Right postcentral gyri grey matter volume harmonized with combat

rh_posteriorcingulate_thickness_combat

Right posterior cingulate thickness harmonized with combat

rh_posteriorcingulate_volume_combat

Right posterior cingulate grey matter volume harmonized with combat

rh_precentral_thickness_combat

Right precentral gyri thickness harmonized with combat

rh_precentral_volume_combat

Right precentral gyri grey matter volume harmonized with combat

rh_precuneus_thickness_combat

Right precuneus thickness harmonized with combat

rh_precuneus_volume_combat

Right precuneus grey matter volume harmonized with combat

rh_rostralanteriorcingulate_thickness_combat

Right rostral anterior cingulate thickness harmonized with combat

rh_rostralanteriorcingulate_volume_combat

Right rostral anterior cingulate grey matter volume harmonized with combat

rh_rostralmiddlefrontal_thickness_combat

Right rostral middle frontal gyri thickness harmonized with combat

rh_rostralmiddlefrontal_volume_combat

Right rostral middle frontal gyri grey matter volume harmonized with combat

rh_superiorfrontal_thickness_combat

Right superior frontal gyri thickness harmonized with combat

rh_superiorfrontal_volume_combat

Right superior frontal gyri grey matter volume harmonized with combat

rh_superiorparietal_thickness_combat

Right superior parietal gyri thickness harmonized with combat

rh_superiorparietal_volume_combat

Right superior parietal gyri grey matter volume harmonized with combat

rh_superiortemporal_thickness_combat

Right superior temporal gyri thickness harmonized with combat

rh_superiortemporal_volume_combat

Right superior temporal gyri grey matter volume harmonized with combat

rh_supramarginal_thickness_combat

Right supramarginal gyri thickness harmonized with combat

rh_supramarginal_volume_combat

Right supramarginal gyri grey matter volume harmonized with combat

rh_temporalpole_thickness_combat

Right temporal pole thickness harmonized with combat

rh_temporalpole_volume_combat

Right temporal pole grey matter volume harmonized with combat

rh_transversetemporal_thickness_combat

Right transverse temporal pole thickness harmonized with combat

rh_transversetemporal_volume_combat

Right transverse temporal pole grey matter volume harmonized with combat

rhCerebralWhiteMatterVol_combat

Right cerebral cortex white matter volume harmonized with combat

rhCortexVol_combat

Right cerebral cortex grey matter volume harmonized with combat

Right.Accumbens.area_combat

Right accumbens volume harmonized with combat

Right.Amygdala_combat

Right amygdala volume harmonized with combat

Right.Caudate_combat

Right caudate volume harmonized with combat

Right.Cerebellum.Cortex_combat

Right cerebellar cortex volume harmonized with combat

Right.Cerebellum.White.Matter_combat

Right cerebellar white matter volume harmonized with combat

Right.choroid.plexus_combat

Right choroid plexus volume harmonized with combat

Right.Hippocampus_combat

Right hippocampus volume harmonized with combat

Right.Inf.Lat.Vent_combat

Right inferior lateral ventricle volume harmonized with combat

Right.Lateral.Ventricle_combat

Right lateral ventricle volume harmonized with combat

Right.Pallidum_combat

Right pallidum volume harmonized with combat

Right.Putamen_combat

Right putamen volume harmonized with combat

Right.Thalamus.Proper_combat

Right thalamus volume harmonized with combat

Right.VentralDC_combat

Right ventral DC volume harmonized with combat

SubCortGrayVol_combat

Total volume of the subcortical grey matter harmonized with combat

SupraTentorialVol_combat

Total volume of brain (minus the cerebellum) harmonized with combat

SupraTentorialVolNotVent_combat

Total volume of brain without ventricles (minus the cerebellum) harmonized with combat

SupraTentorialVolNotVentVox_combat

Number of voxels of brain without ventricles (minus the cerebellum) harmonized with combat

TotalGrayVol_combat

Total volume of all grey matter (cortical and subcortical) harmonized with combat

WM.hypointensities_combat

White matter hypointensities harmonized with combat

Source

https://www.naccdata.org/about-nacc-data/.

Examples

if (FALSE) { # \dontrun{
browseVignettes('NACCADRC')
} # }